+++ js/data.js
@@ -1,3 +1,5 @@
+import {randomNumber} from './utils.js';
+
 const OBJECT = 25;
 
 const NAME = [
@@ -24,16 +26,6 @@ const descriptions = [
 ];
 
 const arrayObjects = [];
-
-const randomNumber =(min, max) => {
-  if (min < 0 || max < 0) {
-    return -1;
-  }
-  if (min > max) {
-    [min, max] = [max, min];
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
 
 const arrayComments = (count) => {
   const array = [];
