+++ js/utils.js
@@ -1,0 +1,11 @@
+const randomNumber =(min, max) => {
+  if (min < 0 || max < 0) {
+    return -1;
+  }
+  if (min > max) {
+    [min, max] = [max, min];
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export {randomNumber};
